% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa.R
\name{calculate_tsa_run_rate}
\alias{calculate_tsa_run_rate}
\title{Calculate TSA passenger check in run rates}
\usage{
calculate_tsa_run_rate(start_year = 2020, end_year = 2025)
}
\arguments{
\item{start_year}{start year of analysis}

\item{end_year}{end year of analysis}
}
\value{
a list of data and ggplot
}
\description{
Downloads tsa passenger check in data and will calculate an annualized number based on year to date
run rates and will also plot the run rates.

A list of data and plot are returned
}
\examples{
calculate_tsa_run_rate(start_year=2020,end_year=2025)
}
