% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-md.R
\name{connect_to_motherduck}
\alias{connect_to_motherduck}
\title{Create connection to motherduck}
\usage{
connect_to_motherduck(motherduck_token = "MOTHERDUCK_TOKEN")
}
\arguments{
\item{motherduck_token}{motherduck token saved in your environment file}
}
\value{
connection
}
\description{
Create connection to motherduck
}
