% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa.R
\name{download_tsa_daily_volumes}
\alias{download_tsa_daily_volumes}
\title{Download TSA Daily Passenger Volumes}
\usage{
download_tsa_daily_volumes()
}
\value{
a tibble
}
\description{
Pulls the daily TSA passenger check-in volumes from tsa.gov website
from 2020 to the current date
}
