% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa.R
\name{download_tsa_daily_volumes_and_upload_to_md}
\alias{download_tsa_daily_volumes_and_upload_to_md}
\title{Download TSA Daily Passenger Volumes}
\usage{
download_tsa_daily_volumes_and_upload_to_md()
}
\value{
invisible
}
\description{
Pulls the daily TSA passenger check-in volumes from tsa.gov website
from 2020 to the current date
}
