% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-etl.R
\name{download_wiki_airports}
\alias{download_wiki_airports}
\title{Download wiki aiports}
\usage{
download_wiki_airports()
}
\value{
tibble
}
\description{
Downloads the list of airports in the USA from the wikipedia page from
'https://en.wikipedia.org/wiki/List_of_airports_in_the_United_States'
This is used in further transformation and cleaning steps
}
